Visualization Critique 3
<h3> Visualization Critique 3 - Oliver Mika</h3>
<h3>Blame the Weather</h3>
<a
href='https://public.tableau.com/s/gallery/blame-weather-us-flight-delayed-precipitation'>
</a>

<p>
For my latest visualization critique, I decided to use one of the
resources provided in the syllabus. This is a serios of maps showing air
travel affected by rain at several airports across the Unites States.
Coming from Seattle, I found this map amusing because i had no idea it
was possible to delay a flight due to rainfall. The map is measured
almost as a cloropleth. There is a Weather Delay Minutes per inch of
Precipitation (WDMIP) score which is measured my color to illustrate how
many or few delays affect each airport.</p>

<p>Visually speaking, i think this map does a pretty good job of getting
its point across. It is legible, easy to understand, and is a topic that
is very precise, but whose data can be used for a variety of different
studies. Color coding helps represent the data well enough to let the
viewer know that the airports in wetter climates actually have fewer
delays than those in drier ones! Also, the colors and design make the
map(s) visually compelling so they are also pleasant to look at.</P>
</div>